change scratchpad_xyhw from xyhw.rs to window_hanler.rs
